feat: add upload carousel doc
