+++ packages/component/src/Upload/demos/basic.tsx
@@ -1,41 +1,6 @@
 import React, { useState } from 'react';
-import { Upload, UploadButton } from 'aelf-design';
-import { UploadProps, UploadRequestOption } from 'antd';
-
-// import type { UploadRequestOption } from 'rc-upload/lib/interface';
-
-import { useAWSUploadService } from '../../hooks/useAws';
-
-const AWSUpload = () => {
-  const [fileList, setFileList] = useState<any[]>([]);
-  const { awsUploadFile } = useAWSUploadService();
-  const handleChange: UploadProps['onChange'] = (info) => {
-    let newFileList = [...info.fileList];
-    newFileList = newFileList
-      .map((file) => {
-        if (file.response) {
-          // Component will show file.url as link
-          file.url = file.response.url;
-        }
-        return file;
-      })
-      .filter((item) => item.status !== 'error');
-    setFileList([...newFileList]);
-  };
-  const customUpload = async ({ file, onSuccess, onError }: UploadRequestOption) => {
-    try {
-      const uploadFile = await awsUploadFile(file as File);
-      console.log(uploadFile, 'uploadFile');
-      onSuccess?.({
-        url: uploadFile,
-      });
-    } catch (error) {
-      onError?.(error as Error);
-    }
-  };
-
-  return <Upload fileList={fileList} customRequest={customUpload} onChange={handleChange} />;
-};
+import { Upload } from 'aelf-design';
+import { UploadProps } from 'antd';
 
 const FileUpload = () => {
   const [fileList, setFileList] = useState<any[]>([]);
@@ -53,22 +18,10 @@ const FileUpload = () => {
   );
 };
 
-const UploadExample: React.FC = () => (
+const App: React.FC = () => (
   <div>
-    <AWSUpload />
     <FileUpload></FileUpload>
-    <UploadButton />
   </div>
 );
 
-export default UploadExample;
-
-// const App: React.FC = () => {
-//   return (
-//     <AELFDProvider prefixCls="zhaomeng">
-//       <Address address={'3ea2cfd153b8d8505097b81c87c11f5d05097c18'} />;
-//     </AELFDProvider>
-//   );
-// };
-
-// export default App;
+export default App;
